+++ spyrmsd/optional/__init__.py
@@ -1,7 +1,0 @@
-# Handle versioneer
-from spyrmsd._version import get_versions
-
-versions = get_versions()
-__version__ = versions["version"]
-__git_revision__ = versions["full-revisionid"]
-del get_versions, versions